move pkg/scheduler/server to cmd/kube-scheduler/app
